requirements.txt
fastapi
sqlalchemy
python-dotenv
pydantic
python-jose
passlib[bcrypt]
langchain-core
langchain-openai
langchain-community
langchain-huggingface
langchain-chroma
unstructured
requests
huggingface-hub
chromadb
psycopg2-binary
python-multipart

